requirements.txt
numpy==1.24.4
matplotlib==3.7.5
scikit-learn==1.3.2
opencv-python>=4.10.0
pillow==10.4.0
albumentations==1.4.18
tqdm==4.67.1
transformers==4.46.3
segmentation-models-pytorch==0.3.3